refactor: extract project resource discovery
